Create Home_For_Chef.java
Our hardworking chef is bored of sleeping in his restaurants. He has decided to settle down. The first thing he must do is to find a suitable location to build a palatial home.

Think of the city as a two-dimensional grid. There are N restaurants in the city. Each of the chef's restaurant is a point denoted by (X , Y). A house can be located at a grid point (R, S) if the sum of the distances between this point and each of the restaurants is as small as possible. Find the number of possible house locations in the city to help out chef build a home.

More than one restaurant can be located at the same point. Houses and restaurants can be located at the same point. Every house must have integer co-ordinates. In other words, R and S are integers. The distance between two points (A,B) and (C,D) is |A-C| + |B-D|. Here |X| is the absolute function.

Difficulty Rating:1815 !